Fixes #415 and #284
In your common class the methods build_paths() and load_settings() are very intertwined.
This results in the force_en-US setting to be ignored after extraction making it stuck on "Installing".
This approach reload the paths if this setting is found.
I'm not really happy with how the fix works but it does.

Hope it helps!